Move async_digest module to separate file
